requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
rembg==2.0.67
Pillow==10.0.1
onnxruntime==1.22.1
gunicorn==21.2.0
